♻️ Refactor: auth 공통 컴포넌트 폴더 분리 (#189)
