+++ src/components/auth/common/AuthBadge.tsx
