+++ src/components/auth/common/AuthIcon.tsx
