+++ src/components/auth/common/AuthLogo.tsx
