+++ src/components/auth/common/AuthStep.tsx
