+++ src/components/auth/common/index.ts
@@ -1,0 +1,25 @@
+import { AuthContainer } from '@/components/auth/common/AuthContainer'
+import { AuthLogo } from '@/components/auth/common/AuthLogo'
+import { AuthTitle } from '@/components/auth/common/AuthTitle'
+import { AuthDescription } from '@/components/auth/common/AuthDescription'
+import { AuthLink } from '@/components/auth/common/AuthLink'
+import { AuthSubmitButton } from '@/components/auth/common/AuthSubmitButton'
+import { AuthVerifyButton } from '@/components/auth/common/AuthVerifyButton'
+import { AuthSocialLoginButton } from '@/components/auth/common/AuthSocialLoginButton'
+import { AuthBadge } from '@/components/auth/common/AuthBadge'
+import AuthIcon from '@/components/auth/common/AuthIcon'
+import AuthStep from '@/components/auth/common/AuthStep'
+
+export {
+  AuthContainer,
+  AuthLogo,
+  AuthTitle,
+  AuthDescription,
+  AuthLink,
+  AuthSubmitButton,
+  AuthVerifyButton,
+  AuthSocialLoginButton,
+  AuthBadge,
+  AuthIcon,
+  AuthStep,
+}
