--- src/components/auth/index.ts
@@ -1,21 +1,0 @@
-import { AuthContainer } from '@/components/auth/AuthContainer'
-import { AuthLogo } from '@/components/auth/AuthLogo'
-import { AuthTitle } from '@/components/auth/AuthTitle'
-import { AuthDescription } from '@/components/auth/AuthDescription'
-import { AuthLink } from '@/components/auth/AuthLink'
-import { AuthSubmitButton } from '@/components/auth/AuthSubmitButton'
-import { AuthVerifyButton } from '@/components/auth/AuthVerifyButton'
-import { AuthSocialLoginButton } from '@/components/auth/AuthSocialLoginButton'
-import { AuthBadge } from '@/components/auth/AuthBadge'
-
-export {
-  AuthContainer,
-  AuthLogo,
-  AuthTitle,
-  AuthDescription,
-  AuthLink,
-  AuthSubmitButton,
-  AuthVerifyButton,
-  AuthSocialLoginButton,
-  AuthBadge,
-}
